--- src/components/Head.js
@@ -1,23 +1,0 @@
-import React from 'react'
-import { Helmet } from 'react-helmet'
-import { useStaticQuery, graphql } from 'gatsby'
-
-const Head = ({ title }) => {
-  const data = useStaticQuery(graphql`
-    query {
-      site {
-        siteMetadata {
-          title
-        }
-      }
-    }
-  `)
-
-  return (
-    <>
-      <Helmet title={`${data.site.siteMetadata.title} | ${title}`} />
-    </>
-  )
-}
-
-export default Head
