+++ src/pages/index.js
@@ -2,12 +2,12 @@ import React from 'react'
 
 import Page from '../components/Page/Page'
 import FrontPage from '../components/FrontPage/FrontPage'
-import Head from '../components/Head'
+import SEO from '../components/SEO'
 
 export default () => {
   return (
     <Page>
-      <Head title="Home" />
+      <SEO title="Home" />
       <FrontPage />
     </Page>
   )
